rename javascript.py to datatables_repr.py
