+++ itables/__init__.py
@@ -1,10 +1,5 @@
-from .javascript import (
-    JavascriptCode,
-    JavascriptFunction,
-    init_notebook_mode,
-    show,
-    to_html_datatable,
-)
+from .datatables_repr import init_notebook_mode, show, to_html_datatable
+from .javascript import JavascriptCode, JavascriptFunction
 from .version import __version__
 
 __all__ = [
